fix -Wshadow and -Wimplicit-int-conversion warnings

+++ src/crt/ulltof.c
@@ -4,7 +4,7 @@
 
 float _ulltof_c(unsigned long long x)
 {
-    uint8_t exponent = x ? __builtin_clzll(x) : ULLONG_WIDTH;
+    uint8_t exponent = x ? ((uint8_t)__builtin_clzll(x)) : ULLONG_WIDTH;
     if (exponent >= ULLONG_WIDTH - ULONG_WIDTH) {
         return (float)((unsigned long)x);
     }
